+++ src/nvim/buffer.c
@@ -76,6 +76,7 @@
 #include "nvim/syntax.h"
 #include "nvim/ui.h"
 #include "nvim/undo.h"
+#include "nvim/usercmd.h"
 #include "nvim/version.h"
 #include "nvim/vim.h"
 #include "nvim/window.h"
