+++ src/nvim/lua/executor.h
@@ -8,9 +8,9 @@
 #include "nvim/assert.h"
 #include "nvim/eval/typval.h"
 #include "nvim/ex_cmds_defs.h"
-#include "nvim/ex_docmd.h"
 #include "nvim/func_attr.h"
 #include "nvim/lua/converter.h"
+#include "nvim/usercmd.h"
 
 // Generated by msgpack-gen.lua
 void nlua_add_api_functions(lua_State *lstate) REAL_FATTR_NONNULL_ALL;
